backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
azure-cognitiveservices-speech==1.31.0
python-dotenv==1.0.0
pydantic>=2.0.0
requests==2.31.0
cohere==4.37 